Analytics and Restart Level Fix
One of the recent updates earlier today caused the metrics UserId and HighestCompletedLevel to not work properly. This fixes that issue. It also allows the user to restart the same level when they die instead of going back to level one, as per the prompt text. ZoomSpeed in CameraController is commented out since it isn't currently used.