Additional Analytics and Graph Generation
Added two more analytics to measure user engagement with level 1. The first one is a stopwatch that tracks how long it takes for the user to complete the first level. The second tracks how often the user died in level 1 before completing it.

Additionally, this commit includes the python scripts for graph generation of analytics using python/matplotlib.