Added new analytics to levels 2 and 3
Time spent and times retried previously only worked for level 1. This expands those metrics to include the additional levels. It also changes the representation of the current level to an integer (instead of a string of the level name) to be used for loading future levels, per the lecture recommendation.